catch up with develop

+++ cgogn/core/types/container/chunk_array.h
@@ -1,25 +1,25 @@
 /*******************************************************************************
-* CGoGN: Combinatorial and Geometric modeling with Generic N-dimensional Maps  *
-* Copyright (C) 2015, IGG Group, ICube, University of Strasbourg, France       *
-*                                                                              *
-* This library is free software; you can redistribute it and/or modify it      *
-* under the terms of the GNU Lesser General Public License as published by the *
-* Free Software Foundation; either version 2.1 of the License, or (at your     *
-* option) any later version.                                                   *
-*                                                                              *
-* This library is distributed in the hope that it will be useful, but WITHOUT  *
-* ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or        *
-* FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License  *
-* for more details.                                                            *
-*                                                                              *
-* You should have received a copy of the GNU Lesser General Public License     *
-* along with this library; if not, write to the Free Software Foundation,      *
-* Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA.           *
-*                                                                              *
-* Web site: http://cgogn.unistra.fr/                                           *
-* Contact information: cgogn@unistra.fr                                        *
-*                                                                              *
-*******************************************************************************/
+ * CGoGN: Combinatorial and Geometric modeling with Generic N-dimensional Maps  *
+ * Copyright (C), IGG Group, ICube, University of Strasbourg, France            *
+ *                                                                              *
+ * This library is free software; you can redistribute it and/or modify it      *
+ * under the terms of the GNU Lesser General Public License as published by the *
+ * Free Software Foundation; either version 2.1 of the License, or (at your     *
+ * option) any later version.                                                   *
+ *                                                                              *
+ * This library is distributed in the hope that it will be useful, but WITHOUT  *
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or        *
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License  *
+ * for more details.                                                            *
+ *                                                                              *
+ * You should have received a copy of the GNU Lesser General Public License     *
+ * along with this library; if not, write to the Free Software Foundation,      *
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301 USA.           *
+ *                                                                              *
+ * Web site: http://cgogn.unistra.fr/                                           *
+ * Contact information: cgogn@unistra.fr                                        *
+ *                                                                              *
+ *******************************************************************************/
 
 #ifndef CGOGN_CORE_CONTAINER_CHUNK_ARRAY_H_
 #define CGOGN_CORE_CONTAINER_CHUNK_ARRAY_H_
@@ -30,9 +30,9 @@
 
 #include <cgogn/core/types/container/attribute_container.h>
 
-#include <vector>
-#include <string>
 #include <memory>
+#include <string>
+#include <vector>
 
 namespace cgogn
 {
@@ -45,11 +45,9 @@ template <typename T>
 class CGOGN_CORE_EXPORT ChunkArray : public AttributeGenT
 {
 public:
-
 	static const uint32 CHUNK_SIZE = 1024;
 
 private:
-
 	std::vector<T*> chunks_;
 	uint32 capacity_;
 
@@ -58,12 +56,11 @@ private:
 		while (index >= capacity_)
 		{
 			chunks_.push_back(new T[CHUNK_SIZE]());
-			capacity_ = chunks_.size() * CHUNK_SIZE;
+			capacity_ = uint32(chunks_.size()) * CHUNK_SIZE;
 		}
 	}
 
 public:
-
 	ChunkArray(AttributeContainerGen* container, const std::string& name) : AttributeGenT(container, name)
 	{
 		chunks_.reserve(512u);
@@ -103,19 +100,19 @@ public:
 	inline void copy(ChunkArray<T>* ca)
 	{
 		if (ca->container_ == this->container_)
-			for (uint32 i = 0; i < chunks_.size(); ++i)
+			for (uint32 i = 0; i < uint32(chunks_.size()); ++i)
 				std::copy(ca->chunks_[i], ca->chunks_[i] + CHUNK_SIZE, chunks_[i]);
 	}
 
 	inline uint32 nb_chunks() const
 	{
-		return chunks_.size();
+		return uint32(chunks_.size());
 	}
 
 	inline std::vector<const void*> chunk_pointers() const
 	{
 		std::vector<const void*> pointers;
-		pointers.reserve(chunks_.size());
+		pointers.reserve(uint32(chunks_.size()));
 		for (auto chunk : chunks_)
 			pointers.push_back(chunk);
 
@@ -128,11 +125,12 @@ public:
 		uint32 index_;
 
 	public:
-
 		inline const_iterator(const ChunkArray<T>* ca, uint32 index) : ca_(ca), index_(index)
-		{}
+		{
+		}
 		inline const_iterator(const const_iterator& it) : ca_(it.ca_), index_(it.index_)
-		{}
+		{
+		}
 		inline const_iterator& operator=(const const_iterator& it)
 		{
 			ca_ = it.ca_;
@@ -153,10 +151,19 @@ public:
 		{
 			return ca_->operator[](index_);
 		}
-		inline uint32 index() { return index_; }
+		inline uint32 index()
+		{
+			return index_;
+		}
 	};
-	inline const_iterator begin() const { return const_iterator(this, this->container_->first_index()); }
-	inline const_iterator end() const { return const_iterator(this, this->container_->last_index()); }
+	inline const_iterator begin() const
+	{
+		return const_iterator(this, this->container_->first_index());
+	}
+	inline const_iterator end() const
+	{
+		return const_iterator(this, this->container_->last_index());
+	}
 
 	class iterator
 	{
@@ -164,11 +171,12 @@ public:
 		uint32 index_;
 
 	public:
-
 		inline iterator(ChunkArray<T>* ca, uint32 index) : ca_(ca), index_(index)
-		{}
+		{
+		}
 		inline iterator(const iterator& it) : ca_(it.ca_), index_(it.index_)
-		{}
+		{
+		}
 		inline iterator& operator=(const iterator& it)
 		{
 			ca_ = it.ca_;
@@ -189,10 +197,19 @@ public:
 		{
 			return ca_->operator[](index_);
 		}
-		inline uint32 index() { return index_; }
+		inline uint32 index()
+		{
+			return index_;
+		}
 	};
-	inline iterator begin() { return iterator(this, this->container_->first_index()); }
-	inline iterator end() { return iterator(this, this->container_->last_index()); }
+	inline iterator begin()
+	{
+		return iterator(this, this->container_->first_index());
+	}
+	inline iterator end()
+	{
+		return iterator(this, this->container_->last_index());
+	}
 };
 
 } // namespace cgogn
